+++ ClubDoorman/Services/ChannelModeration/ChannelModerationModule.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace ClubDoorman.Services.ChannelModeration;
+
+public static class ChannelModerationModule
+{
+    public static IServiceCollection AddChannelModerationServices(this IServiceCollection services)
+    {
+        // TODO: Добавить регистрацию сервисов если необходимо
+        return services;
+    }
+}
